refactor: reduce bundle size by generating design token values at build time
